Practica 2S_2021/requirements.txt
alembic==1.3.2
aniso8601==8.0.0
Click==7.0
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
itsdangerous==1.1.0
Jinja2==2.11.3
Mako==1.1.0
MarkupSafe==1.1.1
psycopg2-binary==2.8.4
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2019.3
six==1.13.0
Werkzeug==0.16.0